Move public lib files into new include tree structure, Move public include files to CRLF EoL sequence wor Windows compat, Redo binding with pybind11 files
