+++ CMakeLists.txt
@@ -95,7 +95,7 @@ if(NOT CMAKE_MODULES_DIR)
   set(CMAKE_MODULES_DIR lib${LIB_SUFFIX}/cmake)
 endif(NOT CMAKE_MODULES_DIR)
 
-set(GR_INCLUDE_DIR      include/bokehgui)
+set(GR_INCLUDE_DIR      include/gnuradio/bokehgui)
 set(GR_CMAKE_DIR        ${CMAKE_MODULES_DIR}/bokehgui)
 set(GR_PKG_DATA_DIR     ${GR_DATA_DIR}/${CMAKE_PROJECT_NAME})
 set(GR_PKG_DOC_DIR      ${GR_DOC_DIR}/${CMAKE_PROJECT_NAME})
@@ -173,7 +173,7 @@ add_custom_target(uninstall
 ########################################################################
 # Add subdirectories
 ########################################################################
-add_subdirectory(include/bokehgui)
+add_subdirectory(include/gnuradio/bokehgui)
 add_subdirectory(lib)
 add_subdirectory(apps)
 add_subdirectory(docs)
